move constants file and fix problem with static path
