+++ inginious/frontend/plugins/manual_scoring/pages/students_list.py
@@ -7,7 +7,7 @@
 
 from collections import OrderedDict
 from inginious.frontend.pages.course_admin.utils import INGIniousAdminPage
-from inginious.frontend.plugins.manual_scoring.pages.constants import get_render_path, get_use_minify
+from inginious.frontend.plugins.manual_scoring.constants import get_render_path, get_use_minify
 
 base_renderer_path = get_render_path()
